refactor: extract component helpers
