+++ apps/admin/src/components/GraphCanvas.tsx
@@ -1,24 +1,5 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-
-export type GraphNode = {
-  key: string;
-  title: string;
-  type?: "start" | "normal" | "end";
-};
-
-export type GraphEdge = {
-  from_node_key: string;
-  to_node_key: string;
-  label?: string | null;
-};
-
-export interface GraphCanvasProps {
-  nodes: GraphNode[];
-  edges: GraphEdge[];
-  onNodeDoubleClick?: (nodeKey: string) => void;
-  onCreateEdge?: (fromKey: string, toKey: string) => void;
-  height?: number; // px
-}
+import type { GraphNode, GraphEdge, GraphCanvasProps } from "./GraphCanvas.helpers";
 
 // Простой layered-лейаут: уровни по расстоянию от стартового узла, fallback — топологическая раскладка
 function computeLayout(nodes: GraphNode[], edges: GraphEdge[]) {
