+++ apps/admin/src/components/ToastProvider.helpers.ts
@@ -1,0 +1,7 @@
+export interface Toast {
+  id: string;
+  title: string;
+  description?: string;
+  variant?: "success" | "error" | "info" | "warning";
+  duration?: number; // ms
+}
