+++ apps/admin/src/components/content/ContentEditor.helpers.ts
@@ -1,0 +1,10 @@
+export const EDITOR_TABS = [
+  "General",
+  "Content",
+  "Relations",
+  "AI",
+  "Validation",
+  "History",
+  "Publishing",
+  "Notifications",
+];
