+++ apps/admin/src/components/content/GeneralTab.tsx
@@ -2,17 +2,7 @@ import FieldTitle from "../fields/FieldTitle";
 import FieldSlug from "../fields/FieldSlug";
 import FieldTags from "../fields/FieldTags";
 import FieldCover from "../fields/FieldCover";
-
-export interface GeneralTabProps {
-  title: string;
-  slug: string;
-  tags: string[];
-  cover: string | null;
-  onTitleChange: (v: string) => void;
-  onSlugChange: (v: string) => void;
-  onTagsChange: (t: string[]) => void;
-  onCoverChange: (url: string | null) => void;
-}
+import type { GeneralTabProps } from "./GeneralTab.helpers";
 
 export default function GeneralTab({
   title,
