+++ apps/admin/src/pages/QuestVersionEditor.tsx
@@ -2,7 +2,8 @@ import { useEffect, useMemo, useState } from "react";
 import { useParams } from "react-router-dom";
 import { getVersion, putGraph, publishVersion, validateVersion, autofixVersion, type VersionGraph } from "../api/questEditor";
 import PageLayout from "./_shared/PageLayout";
-import NodeEditorModal, { type NodeEditorData } from "../components/NodeEditorModal";
+import NodeEditorModal from "../components/NodeEditorModal";
+import type { NodeEditorData } from "../components/NodeEditorModal.helpers";
 import MediaPicker from "../components/MediaPicker";
 import { getQuestMeta, updateQuestMeta } from "../api/questEditor";
 import GraphCanvas from "../components/GraphCanvas";
